Fixes to merge issues
git-svn-id: https://macademia.googlecode.com/svn/branches/nbrviz@802 22e0cbe0-2fb8-11df-a735-6ff41f51f3aa

+++ Macademia/grails-app/services/org/macademia/nbrviz/PseudonymService.groovy
@@ -61,7 +61,7 @@ class PseudonymService {
         def fakeData = [:]
         fakeData['id']=pair(sid, uid)
         fakeData['name']=generatePseudonym((long) fakeData.id)
-        fakeData['pic']=generateFakePicture(fakeData.id)
+        fakeData['pic']=generateFakePicture(fakeData.id as Long)
         return fakeData
     }
